requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pandas>=2.0.0
numpy>=1.24.0
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
groq>=0.4.0
httpx>=0.25.0
python-multipart>=0.0.6
motor>=3.3.0
pymongo>=4.6.0
dnspython>=2.4.0
datasets>=2.14.0
huggingface-hub>=0.16.0